new commit

+++ Rythm game/src/Beatmap/Note.java
@@ -1,5 +1,5 @@
 package Beatmap;
 
 public interface Note {
-	public int getTime (); // where note need to be hit
+	public float getTime (); // where note need to be hit
 }
